+++ src/lib/runtime/Web3FunctionRunner.ts
@@ -193,7 +193,9 @@ export class Web3FunctionRunner {
       }
       storage = {
         ...data.storage,
-        size: Buffer.byteLength(JSON.stringify(data.storage), "utf-8") / 1024,
+        size:
+          Buffer.byteLength(JSON.stringify(data.storage.storage), "utf-8") /
+          1024,
       };
     } catch (err) {
       error = err;
@@ -327,9 +329,13 @@ export class Web3FunctionRunner {
         `Web3Function exited with code=${signal} (RPC requests limit exceeded)`,
         "rpcRequest"
       );
+    } else if (signal === 251) {
+      return new Web3FunctionRuntimeError(
+        `Web3Function exited with code=${signal} (Unhandled promise rejection)`
+      );
     } else if (
       (runtime === "docker" && signal === 137) ||
-      (runtime === "thread" && this._memory >= memoryLimit)
+      (runtime === "thread" && this._memory / 1024 / 1024 >= memoryLimit)
     ) {
       return new Web3FunctionRuntimeError(
         `Web3Function exited with code=${signal} (Memory limit exceeded)`,