+++ src/lib/types/Web3FunctionContext.ts
@@ -53,6 +53,8 @@ export interface Web3FunctionContext {
     get(key: string): Promise<string | undefined>;
     set(key: string, value: string): Promise<void>;
     delete(key: string): Promise<void>;
+    getKeys(): Promise<string[]>;
+    getSize(): Promise<number>;
   };
 }